[refact] slowapi > 경고 없이 사용할 Limiter를 범용성을 위해 상위경로로 이동
